requirements.txt
drake
protobuf
grpcio-tools
jupyter
pyserial
varint
dacite
psutil
saleae
aiohttp
pandas
